+++ src/index/quadtree/DoubleBits.cpp
@@ -8,7 +8,7 @@
  *
  * This is free software; you can redistribute and/or modify it under
  * the terms of the GNU Lesser General Public Licence as published
- * by the Free Software Foundation. 
+ * by the Free Software Foundation.
  * See the COPYING file for more information.
  *
  **********************************************************************
@@ -118,7 +118,7 @@ int
 DoubleBits::getExponent() const
 {
 #if ASSUME_IEEE_DOUBLE
-	return biasedExponent()-EXPONENT_BIAS;
+	return static_cast<int>(biasedExponent() - EXPONENT_BIAS);
 #else
 	if ( x <= 0 ) return 0; // EDOM || ERANGE
 	return (int)((log(x)/log(2.0))+(x<1?-0.9:0.00000000001));
