+++ src/noding/NodingValidator.cpp
@@ -8,10 +8,10 @@
  *
  * This is free software; you can redistribute and/or modify it under
  * the terms of the GNU Lesser General Licence as published
- * by the Free Software Foundation. 
+ * by the Free Software Foundation.
  * See the COPYING file for more information.
  *
- ********************************************************************** 
+ **********************************************************************
  *
  * Last port: noding/NodingValidator.java rev. 1.6 (JTS-1.7)
  *
@@ -36,7 +36,7 @@ namespace noding { // geos.noding
 
 /*public*/
 void
-NodingValidator::checkValid() 
+NodingValidator::checkValid()
 {
 	checkEndPtVertexIntersections();
 	checkInteriorIntersections();
@@ -62,7 +62,7 @@ void
 NodingValidator::checkCollapses(const SegmentString& ss) const
 {
 	const CoordinateSequence& pts = *(ss.getCoordinates());
-	for (unsigned int i=0, n=pts.getSize()-2; i<n; ++i)
+	for (unsigned int i=0, n = static_cast<unsigned int>(pts.getSize() - 2); i < n; ++i)
 	{
 		checkCollapse(pts[i], pts[i + 1], pts[i + 2]);
 	}
@@ -82,7 +82,7 @@ NodingValidator::checkCollapse(const Coordinate& p0,
 
 /*private*/
 void
-NodingValidator::checkInteriorIntersections() 
+NodingValidator::checkInteriorIntersections()
 {
 	for (SegmentString::NonConstVect::const_iterator
 		it = segStrings.begin(), itEnd = segStrings.end();
@@ -98,18 +98,18 @@ NodingValidator::checkInteriorIntersections()
 			checkInteriorIntersections(*ss0, *ss1);
 		}
 	}
- 
+
 }
 
 /* private */
 void
 NodingValidator::checkInteriorIntersections(const SegmentString& ss0,
-		const SegmentString& ss1) 
+		const SegmentString& ss1)
 {
 	const CoordinateSequence& pts0 = *(ss0.getCoordinates());
 	const CoordinateSequence& pts1 = *(ss1.getCoordinates());
-	for (unsigned int i0=0, n0=pts0.size(); i0<n0-1; i0++) {
-		for (unsigned int i1=0, n1=pts1.size(); i1<n1-1; i1++) {
+	for (unsigned int i0=0, n0 = static_cast<unsigned int>(pts0.size()); i0 < n0 - 1; i0++) {
+		for (unsigned int i1=0, n1 = static_cast<unsigned int>(pts1.size()); i1 < n1 - 1; i1++) {
 			checkInteriorIntersections(ss0, i0, ss1, i1);
 		}
 	}
@@ -120,7 +120,7 @@ NodingValidator::checkInteriorIntersections(const SegmentString& ss0,
 void
 NodingValidator::checkInteriorIntersections(
 		const SegmentString& e0, unsigned int segIndex0,
-		const SegmentString& e1, unsigned int segIndex1) 
+		const SegmentString& e1, unsigned int segIndex1)
 {
 	if (&e0 == &e1 && segIndex0 == segIndex1) return;
 	const Coordinate& p00 = e0.getCoordinates()->getAt(segIndex0);
@@ -171,7 +171,7 @@ NodingValidator::checkEndPtVertexIntersections(const Coordinate& testPt,
 	{
 		const SegmentString* ss0 = *it;
 		const CoordinateSequence& pts = *(ss0->getCoordinates());
-		for (unsigned int j=1, n=pts.size()-1; j<n; ++j)
+		for (unsigned int j=1, n=static_cast<unsigned int>(pts.size()-1); j < n; ++j)
 		{
 			if (pts[j].equals(testPt))
 			{
@@ -184,9 +184,9 @@ NodingValidator::checkEndPtVertexIntersections(const Coordinate& testPt,
 	}
 }
 
- 
 
- 
+
+
 
 /* private */
 bool
